+++ examples/petstore/client.rb
@@ -8,8 +8,8 @@ module Petstore
     server "https://petstore3.swagger.io/api/v3"
     security({ petstore_auth: :bearer, api_key: { header: "api_key" } })
 
-    put    :update_pet, "/pet", body: Pet, into: Pet, security: :petstore_auth
-    post   :add_pet, "/pet", body: Pet, into: Pet, security: :petstore_auth
+    put    :update_pet, "/pet", body: Pet, body_required: true, into: Pet, security: :petstore_auth
+    post   :add_pet, "/pet", body: Pet, body_required: true, into: Pet, security: :petstore_auth
     get    :find_pets_by_status, "/pet/findByStatus", query: %i[status], required: %i[status], into: [Pet], security: :petstore_auth
     get    :find_pets_by_tags, "/pet/findByTags", query: %i[tags], required: %i[tags], into: [Pet], security: :petstore_auth
     get    :get_pet_by_id, "/pet/{petId}", into: Pet, security: [[:api_key], [:petstore_auth]]